# src/api/swagger.json
{
  "openapi": "3.0.0",
  "servers": [],
  "info": {
    "title": "",
    "description": "",
    "version": ""
  },
...
  "paths": {
...
    "/shops": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "List all shops",
        "description": "Returns all shops",
        "operationId": "FindAllShops",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Shop"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Shop"
        ],
        "summary": "Create new shop",
        "description": "",
        "operationId": "CreateShop",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Shop"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewShop"
        }
      }
    },
    "/shops/{id}": {
      "get": {
        "tags": [
          "Shop"
        ],
        "summary": "Get shop",
        "description": "Returns the given shop",
        "operationId": "FindShop",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of shop to return",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Shop"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Shop"
        ],
        "summary": "Update shop",
        "description": "Updates the given shop",
        "operationId": "UpdateShop",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of shop to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Shop"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewShop"
        }
      },
      "delete": {
        "tags": [
          "Shop"
        ],
        "summary": "Delete shop",
        "description": "Removes the given shop",
        "operationId": "DeleteShop",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of shop to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "List all tags",
        "description": "Returns all tags",
        "operationId": "FindAllTags",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tag"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tag"
        ],
        "summary": "Create new tag",
        "description": "",
        "operationId": "CreateTag",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Tag"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewTag"
        }
      }
    },
    "/tags/{id}": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Get tag",
        "description": "Returns the given tag",
        "operationId": "FindTag",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of tag to return",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Tag"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tag"
        ],
        "summary": "Update tag",
        "description": "Updates the given tag",
        "operationId": "UpdateTag",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of tag to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Tag"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewTag"
        }
      },
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "Delete tag",
        "description": "Removes the given tag",
        "operationId": "DeleteTag",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of tag to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "requestBodies": {
...
      "NewShop": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NewShop"
            }
          }
        },
        "description": "Shop object that needs to be added to the database",
        "required": true
      },
...
      "NewTag": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NewTag"
            }
          }
        },
        "description": "Tag object that needs to be added to the database",
        "required": true
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "NewShop": {
        "type": "object",
        "properties": {
          "SPName": {
            "type": "string"
          },
          "SPShortName": {
            "type": "string"
          },
          "SPDescription": {
            "type": "string"
          },
          "SPCountry": {
            "type": "string"
          },
          "SPCity": {
            "type": "string"
          },
          "SPAddress": {
            "type": "string"
          },
          "SPWeb": {
            "type": "string"
          },
          "SPWebStore": {
            "type": "string"
          },
          "SPContact": {
            "type": "string"
          },
          "SPRating": {
            "type": "integer"
          }
        }
      },
...
      "NewTag": {
        "type": "object",
        "properties": {
          "TGName": {
            "type": "string"
          }
        }
      },
...
    }
  }
}